Adding better support for a custom good CA.
Now, if the customgoodca config option is defined, `rake certs:ca` will
check whether it exists or not and will generate it or raise on error as
needed. `rake certs:ssl` now calls certs:ca first, and it will only
generate a new goodca if customgoodca does not exist, or if we somehow
got passed certs:ca without verifying the CA.

We don't use the keypass option for anything other than to read an
existing goodca's private key. If we generate a new CA private key, it
won't be encrypted.