Modifying TestManager#test_completed.
Its first argument is now the test that is completed. Using this, it can
report the correct test since a web browser will cause multiple
connections (and thus multiple TestListeners) to happen simultaneously,
but there is still only one TestManager for all of them. Additionally,
we check whether the current_test matches the completed test in order to
avoid skipping over tests when multiple connections complete the same
test.